pyproject.toml
[project]
name = "Rabbit-in-a-Blender"
version = "0.0.36"
authors = [
  { name="Lammertyn Pieter-Jan", email="pieter-jan.lammertyn@azdelta.be" },
  { name="Dupulthys Stijn", email="stijn.dupulthys@azdelta.be" },
  { name="De Jaeger Peter", email="peter.dejaeger@azdelta.be" }
]
description = "An ETL pipeline to transform your EMP data to OMOP."
readme = "README.md"
license = { file="LICENSE" }
keywords= ["OMOP", "CDM", "common data model", "OHDSI"]
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
]
dependencies = [
    'backoff >= 2.2.1',
    'connectorx >= 0.3.2',
    'polars >= 0.19.13',
    'jinja2 >= 3.1.2',
    'pyarrow >= 14.0.1',
    'google-cloud-bigquery >= 3.13.0',
    'google-cloud-storage >= 2.13.0',
    'google-auth >= 2.23.4',
    'simple-ddl-parser >= 0.31.1',
    'pandas >= 2.1.3',
    'humanfriendly >= 10.0',
    'jpype1 >= 1.4.1',
    'dash >= 2.14.1',
    'dash-table >= 5.0.0',
    'dash-bootstrap-components >= 1.5.0'
]

[project.urls]
"Homepage" = "https://radar-azdelta.github.io/Rabbit-in-a-Blender/"
"Source Code" = "https://github.com/RADar-AZDelta/Rabbit-in-a-Blender"
"Bug Tracker" = "https://github.com/RADar-AZDelta/Rabbit-in-a-Blender/issues"

[project.scripts]
riab = "riab:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"riab.etl" = ["**/*.json", "**/*.sql", "**/*.jinja"]
"riab.libs.SqlRender.inst.csv" = ["*.csv"]
"riab.libs.SqlRender.inst.java" = ["*.jar"]
"riab.libs.DataQualityDashboard.inst.csv" = ["*.csv"]
"riab.libs.DataQualityDashboard.inst.sql.sql_server" = ["*.sql"]
"riab.libs.Achilles.inst.csv" = ["**/*.csv"]
"riab.libs.Achilles.inst.sql.sql_server" = ["**/*.sql"]
"riab.assets" = ["*.css"]

[tool.ruff]
line-length = 120